% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataLoad}
\alias{dataLoad}
\title{\code{dataLoad} format raw stream data}
\usage{
dataLoad(filepath, lat, long, tz)
}
\arguments{
\item{filepath}{character vector, file path to data file, including file name and extension}

\item{lat}{character vector, decimal latitude of site location}

\item{long}{character vector, decimal longitude of site location}

\item{tz}{character vector of the time zone, as recognized by OlsonNames}
}
\value{
dataframe of date-time in UTC, discharge in m^3/s, water temperature in C, dissolved oxygen in mg/L and percent saturation, site latitude and longitude
}
\description{
load local datafile (date-time, discharge in ft^3/s, water temperature in C, dissolved oxygen in mg/L and percent saturation), transform local time to UTC time, convert from American to SI units, attach latitude and longitude to dataframe, and then return data to user and ready to pump into \code{\link{dataPrep}}
}
\examples{
dataLoad(filepath = "./input_files/streamdata.csv", lat = "40", long = "-100", tz = "America/Chicago")
}
