% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.R
\name{dataPrep}
\alias{dataPrep}
\title{\code{dataPrep} Interpolate missing data points and estimate PAR}
\usage{
dataPrep(data, na.fill = "interpolation")
}
\arguments{
\item{data}{dataframe, piped from dataLoad function, contains site information and stream data}

\item{na.fill}{character vector, algorithm that will be used by \code{\link[imputeTS]{na.seasplit}} to fill gaps in the time series data. See documentation of \code{\link[imputeTS]{na.seasplit}} for a list of options. Defaults to interpolation}
}
\value{
dataframe, reformatted and ready to pipe into metabolismModeling.R
}
\description{
preps data (that has been formatted by \code{\link{dataLoad}}) for metabolism modeling. Interpolates missing data points according to method of choice, estimates PAR from longitude and time data, estimates stream depth from discharge, converts DO saturation in percent to DO saturation in mg/L
}
\examples{
data <- dataLoad(filepath = "./input_files/streamdata.csv", lat = "40", long = "-100", tz = "America/Chicago")
dataPrep(data, na.fill = "interpolation")
}
