% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolismPlot.R
\name{metabolismPlot}
\alias{metabolismPlot}
\title{\code{metabolismPlot} generate a pretty plot of GPP and ER estimates, with 95% confidence interval bands}
\usage{
metabolismPlot(metab_output, filename)
}
\arguments{
\item{metab_output}{the list object returned by metabolismModeling.R}

\item{filename}{character vector of filename for plot (including file extension)}
}
\value{
returns plot of metabolism data, as described in function overview
}
\description{
returns a formatted plot of metabolism modeling results, including confidence bands on model estimates.
}
\examples{
data <- dataLoad(filepath = "./input_files/streamdata.csv", lat = "40", long = "-100", tz = "America/Chicago")
data <- dataPrep(data, na.fill = "interpolation")
model_data <- metabolismModeling(data, filename = "metabolismmodel_results.csv")
metabolismPlot(model_data, filename = "metabolismplot.png")

}
